finish refactor sync dialog
